refactor show release notes action (#162823)
Related-to: #162713
